New interface TabbedFormModel for activate first tab with error input
