+++ CHANGELOG_DEV.md
@@ -3,4 +3,4 @@
 
 - Fix #5434: Hide disabled next/prev buttons on guide first/last steps
 - Fix #5456: `canImpersonate` only possible for SystemAdmins
-- Enh #5472: New interface `TabbedForm` for activate first tab with error input
+- Enh #5472: New interface `TabbedFormModel` for activate first tab with error input
